Zshbrowser script
When start, it will spawn ranger from zsh browse() function. This way you can
quit ranger, return to the shell, do some shell stuff, and go back to ranger
with the browse() function.

The big advantage of using ranger this way is that the browse() function syncs
the pwd between ranger and Zsh.